routes added, graph bug fixed
