+++ lib/pages/leaderboard_page.dart
@@ -16,17 +16,25 @@ import 'package:flutter/material.dart';
 /// the SocialMediaPage, builds and displays 
 /// social media page view
 ///*********************************
-class SocialMediaPage extends StatelessWidget {
-  const SocialMediaPage({super.key});
+class LeaderBoardPage extends StatelessWidget {
+  const LeaderBoardPage({super.key});
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-      title: const Text("ProcrastiLeaderboards"),
-      ),
-    body: const Center(
-      child: Text("Friend Icons")
-    )
+    return GestureDetector(
+      onHorizontalDragEnd: (details) {
+        if (details.primaryVelocity != null && details.primaryVelocity! > 0) {
+          Navigator.pushNamed(context, '/leaderBoardPageBack');
+        }
+      },
+      child: Scaffold(
+        appBar: AppBar(
+          automaticallyImplyLeading: false,
+          title: const Text("ProcrastiLeaderboards"),
+        ),
+        body: const Center(
+          child: Text("Friend Icons")
+        )
+      )
     );
   }
 }
