+++ lib/profile/auth_gate.dart
@@ -42,6 +42,7 @@ class AuthGate extends StatelessWidget {
         }
         // If user is signed in run App
         return const ProcrastiHater();
+        //Navigator.pushNamed(context, '/homePage');
       },      
     );
   }
